requirements.txt
hyperopt==0.2.7
matplotlib==3.6.0
networkx==2.8.6
numpy==1.23.3
pandas==1.5.0
prettytable==3.4.1
psycopg2_binary==2.9.3
rdflib==6.2.0
scikit_learn==1.1.2
scipy==1.9.1
seaborn==0.12.2
sqlparse==0.3.1
pytz==2021.3
SQLAlchemy==1.4.22
argparse==1.4.0
torch==2.0.1
torch_geometric==2.3.1
torch_scatter==2.1.1
nltk==3.8.1